feat(app): Add UIView extension for convenient way of adding multiple subviews.
